requirements.txt
Django==3.0.2
gunicorn==20.0.4
django-heroku==0.3.1
django-crispy-forms==1.8.1
django-allauth==0.41.0
whitenoise==5.0.1
dj_database_url==0.5.0
psycopg2-binary==2.8.4